feat : MyEventInfo분리
