+++ src/components/common/MyEventCard/MyEventCard.style.ts
@@ -1,4 +1,3 @@
-import Theme from '@/styles/Theme';
 import styled from '@emotion/styled';
 
 const MyEventCardContainer = styled.div`
@@ -11,19 +10,6 @@ const EventLeftSection = styled.div`
   display: flex;
 `;
 
-const MyEventTitleStyled = styled.div`
-  font-size: 1rem;
-  line-height: 1.2rem;
-  margin: 0.8rem 0;
-`;
-
-const EventInfoSection = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-evenly;
-  margin-left: 1rem;
-`;
-
 const EventRightSection = styled.div`
   width: 5rem;
   min-width: 5rem;
@@ -33,28 +19,4 @@ const EventRightSection = styled.div`
   flex-direction: column;
 `;
 
-const SemiBlackFont = styled.div`
-  color: ${Theme.color.semiBlack};
-  bottom: 0;
-  margin: 0.1rem 0 0 0.2rem;
-  font-size: 0.7rem;
-`;
-
-const EventDescription = styled.div`
-  display: flex;
-  align-items: center;
-  position: relative;
-  font-size: 0.8rem;
-  display: flex;
-  margin-bottom: 0.1rem;
-`;
-
-export {
-  MyEventCardContainer,
-  EventLeftSection,
-  EventRightSection,
-  EventInfoSection,
-  SemiBlackFont,
-  EventDescription,
-  MyEventTitleStyled,
-};
+export { MyEventCardContainer, EventLeftSection, EventRightSection };
